Move property decoding classes to separate package
