+++ FitNesseRoot/PlugIns/JdbcSlim/UserGuide/3HowtoConfigureTheDatabaseConnection/EncryptingPasswords.wiki
@@ -8,7 +8,7 @@ Test
 !3  WARNING: Any smart user of FitNesse can decryt the encrypted passwords!
 The current implementation avoids that passwords are stored in clear text in files.
 If you need more security implement a different store for your passwords.
-You can configure a different decrypting mechanism by creating a class implementing `six42.fitnesse.jdbcslim.PropertyDecoder` and setting ".propertyDecoder" to contain its name.
+You can configure a different decrypting mechanism by creating a class implementing `six42.fitnesse.jdbcslim.propertydecode.PropertyDecoder` and setting ".propertyDecoder" to contain its name.
 
 !4 CREDIT: The Encryption implementation is taken from the [[dbfit][http://github.com/dbfit/dbfit]] library.
