+++ build.xml
@@ -63,7 +63,7 @@
     <target name="compile"  description="compile java sources">
         <mkdir dir="${classes.dir}" />
         <javac srcdir="${src.dir}" destdir="${classes.dir}"  debug="true" source="1.6" target="1.6" includeantruntime="false">
-            <include name="six42/fitnesse/jdbcslim/*.java"/>
+            <include name="six42/fitnesse/jdbcslim/**/*.java"/>
 			<classpath refid="jdbcslim-classpath"/>
         </javac>
     </target>
